have the QR code to be shown only when breaking a tamper seal
Making the QR code is extremely slow on android and it make it seem like the app is broken. This is pushing the "bug" experience to be after the splash screen. However it's good to remind that there is a tamper seal.
Will be working on finding a quicker QR code generetor that plays nicely with expo